implement role-policy/gate

+++ packages/core/src/Http/Controllers/Api/Tenant/RoleController.php
@@ -8,21 +8,18 @@ use Froxlor\Core\Http\Requests\Tenant\StoreTenantRoleRequest;
 use Froxlor\Core\Http\Requests\UpdateRoleRequest;
 use Froxlor\Core\Models\Role;
 use Froxlor\Core\Models\Tenant;
-use Froxlor\Core\Services\Traits\TenantAccessPermission;
 use Froxlor\Core\Support\Response;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
-    use TenantAccessPermission;
-
     /**
      * Display a listing of the resource.
      */
     public function index(Request $request, Tenant $tenant)
     {
-        //Gate::authorize('tenantViewAny', [Role::class, $tenant]);
+        Gate::authorize('tenantViewAny', [Role::class, $tenant]);
 
         $roles = Role::query()
             ->with(['permissions'])
@@ -42,7 +39,7 @@ class RoleController extends Controller
      */
     public function store(StoreTenantRoleRequest $request, Tenant $tenant)
     {
-        Gate::authorize('tenantCreate', Role::class);
+        Gate::authorize('tenantCreate', [Role::class, $tenant]);
 
         // get validated data only for ourselves
         $roleData = $request->validatedResource();
@@ -64,7 +61,7 @@ class RoleController extends Controller
      */
     public function show(Request $request, Tenant $tenant, Role $role)
     {
-       // Gate::authorize('tenantView', [$role, $tenant]);
+        Gate::authorize('tenantView', [$role, $tenant]);
 
         return Response::jsonResource($role);
     }
@@ -74,6 +71,8 @@ class RoleController extends Controller
      */
     public function update(UpdateRoleRequest $request, Tenant $tenant, Role $role)
     {
+        Gate::authorize('tenantUpdate', [$role, $tenant]);
+
         $role->update($request->validated());
 
         return Response::jsonResource($role->refresh());
@@ -84,6 +83,8 @@ class RoleController extends Controller
      */
     public function destroy(Request $request, Tenant $tenant, Role $role)
     {
+        Gate::authorize('tenantDelete', [$role, $tenant]);
+
         $role->delete();
 
         return response()->noContent();
